+++ src/clojure2d/java/reconstruction/Box.java
@@ -6,7 +6,15 @@ public class Box extends AFilter {
         init();
     }
 
+    public Box() {
+        this(0.5);
+    }
+    
     public double evaluate(double x, double y) { 
         return 1.0;
+    }
+
+    public String getName() {
+        return "Box";
     }
 }